Practical 3 - Classification
Classification in machine learning is a supervised learning technique used to categorize data into predefined classes or categories. 

It involves training a model on labeled data (where the correct class is known) to predict the class of new, unseen data points. 

We can think of it as teaching a computer to sort items into different bins based on certain characteristics.