Practical 6 - Logistic Regression
Logistic regression is a supervised machine learning algorithm primarily used for binary classification problems. It models the probability of a binary outcome (e.g., yes/no, true/false, 0/1) based on input features. 
It works by fitting a sigmoid function to the data, which outputs a probability value between 0 and 1. This probability is then used to classify the instance into one of the two classes.