Practical 5 - Linear Regression
Linear regression is a fundamental supervised machine learning algorithm used to model the relationship between a dependent variable (the outcome you want to predict) and one or more independent variables (the features used for prediction) by fitting a linear equation to the observed data. 
Essentially, it aims to find the "best fit" straight line (or hyperplane in higher dimensions) that represents the relationship between these variables.